simgle trips list
